+++ tgbotapi.core/src/commonMain/kotlin/dev/inmo/tgbotapi/types/message/GroupMessages.kt
@@ -1,0 +1,49 @@
+package dev.inmo.tgbotapi.types.message
+
+import dev.inmo.tgbotapi.types.*
+import dev.inmo.tgbotapi.types.buttons.InlineKeyboardMarkup
+import dev.inmo.tgbotapi.types.message.content.abstracts.MessageContent
+import dev.inmo.tgbotapi.types.message.content.abstracts.PossiblySentViaBotCommonMessage
+import dev.inmo.tgbotapi.types.message.payments.SuccessfulPaymentInfo
+import com.soywiz.klock.DateTime
+import dev.inmo.tgbotapi.types.chat.abstracts.*
+import dev.inmo.tgbotapi.types.message.abstracts.*
+
+data class FromChannelGroupMessageImpl<T : MessageContent>(
+    override val chat: GroupChat,
+    override val channel: ChannelChat,
+    override val messageId: MessageIdentifier,
+    override val date: DateTime,
+    override val forwardInfo: ForwardInfo?,
+    override val editDate: DateTime?,
+    override val replyTo: Message?,
+    override val replyMarkup: InlineKeyboardMarkup?,
+    override val content: T,
+    override val senderBot: CommonBot?,
+    override val authorSignature: AuthorSignature?
+) : FromChannelGroupMessage<T>
+
+data class AnonymousGroupMessageImpl<T : MessageContent>(
+    override val chat: GroupChat,
+    override val messageId: MessageIdentifier,
+    override val date: DateTime,
+    override val forwardInfo: ForwardInfo?,
+    override val editDate: DateTime?,
+    override val replyTo: Message?,
+    override val replyMarkup: InlineKeyboardMarkup?,
+    override val content: T,
+    override val senderBot: CommonBot?,
+    override val authorSignature: AuthorSignature?
+) : AnonymousGroupMessage<T>
+
+data class CommonGroupMessageImpl<T : MessageContent>(
+    override val chat: GroupChat,
+    override val messageId: MessageIdentifier,
+    override val date: DateTime,
+    override val forwardInfo: ForwardInfo?,
+    override val editDate: DateTime?,
+    override val replyTo: Message?,
+    override val replyMarkup: InlineKeyboardMarkup?,
+    override val content: T,
+    override val senderBot: CommonBot?
+) : CommonGroupMessage<T>