+++ tgbotapi.core/src/commonMain/kotlin/dev/inmo/tgbotapi/types/message/abstracts/PublicMessage.kt
@@ -1,0 +1,9 @@
+package dev.inmo.tgbotapi.types.message.abstracts
+
+import dev.inmo.tgbotapi.types.chat.abstracts.PublicChat
+import dev.inmo.tgbotapi.types.message.content.abstracts.MessageContent
+import dev.inmo.tgbotapi.types.message.content.abstracts.PossiblySentViaBotCommonMessage
+
+interface PublicMessage<T: MessageContent> : PossiblySentViaBotCommonMessage<T> {
+    override val chat: PublicChat
+}