+++ tgbotapi.core/src/commonMain/kotlin/dev/inmo/tgbotapi/types/message/abstracts/SignedMessage.kt
@@ -1,0 +1,7 @@
+package dev.inmo.tgbotapi.types.message.abstracts
+
+import dev.inmo.tgbotapi.types.AuthorSignature
+
+interface SignedMessage : Message {
+    val authorSignature: AuthorSignature?
+}